Remove the copied .gz file from wd if it exists

+++ R/calmet_get_ncdc_station_data.R
@@ -145,9 +145,8 @@ calmet_get_ncdc_station_data <- function(data_filename = NULL,
           # Extract the downloaded data file
           system("gunzip *.gz", intern = FALSE, ignore.stderr = TRUE)
           
-          # Remove the .gz file from the working folder
-          file.remove(outputs[i, 1])
-          
+          # Remove the .gz file from the working folder if it exists
+          if (file.exists(outputs[i, 1])) file.remove(outputs[i, 1])    
           
         }
